WIP: enabling calls
Some work toward enabling function calls.
For now all calls are marked as non-colocated (ideally functions from
the same module should be marked as colocated), therefore requiring a
jump through an absolute address or an indirection though a table.
The call instructions are emitted, but the relocations are not yet
handled, leading to segmentation fault (by jumping to 0x00).