WIP: handle memory
This commit experiment with heaps, that can be used as Wasm memory. To
do so, the commit introduces the VMContext, a structure pointing to
various data structures owned by an instance (well, only heaps for now).
The VMContext is passed from one function to another as an extra
argument, ideally we could optimize the argument away in some case, but
for now we just add it in every signatures.

This commit is buggy, there is an illegal instruction error that I'll
try to fix later.