Working memory loads & stores
Finilize the heap allocation & register heaps in the module info so that
they are added to the VMContext.

The code is quite crapy for now, but at least that's a first working
version. The heap management can be greatly improved and the
initialization of VMContext can be made much more robust.