WIP: major refactor - everything is broken
Hoist the allocators and instances out of the compiler module.
This is done in preparation for the kernel integration, I plan to make
the compiler integration optionnal (so that the kernel can be compiler
without, and therefore execute only pre-compiled programs).

In the future, there will be (at least) 3 crates:
- Traits: containing the traits and structs that are common to the
  kernel and compiler.
- Compiler: compilers for the various backends. Can be used either AOT
  as a standalone binary or JIT inside of the kernel.
- Kernel: will contain the allocation & module instances logic.