Add support for context switch between modules
Calling a function from an imported module now correctly performs a
context switch by loading the external module's VMContext and passing it
as argument. This enables proper import of functions from external
modules.

The implementation requires knowledge of imported functions and tacking
modules throughout the compilation pipeline. As a consequence there is
now better support for internal calls that are marked as collocated
(and thus called using a relative address).