Handle global imports & exports
Extend globals support to imports & exports.
Imported globals are not directly stored in the current VMContext
(obviously), but are insted pointed to from the corresponding VMContext
entry.

Once again, the VMContext layout does not lend itself well to this kinds
of things. This is not helped by the fact that the layout is fluid (it
depends on the Module implementation) and uniform (all *const u8).
Ideally, a would like to have a VMContext that is more strongly typed,
with methods such as `get_imported_global_ptr` directly on the
VMContext.