VMContext refactor
The VMContext used to be represented as a Vec<*const u8>, which made it
difficult to use it to store vales other than pointers (e.g. globals).
This new version use a custom structure to represent the VMContext
through a raw array of bytes managed manually.
Currently, the array of byte has an alignment of 8 and each item has a
width of 8 bytes, as it was the case before. In the future, the
VMContext can be extended to hold values of different sizes (e.g. 16
bytes for SIMD vectors).