Enable memory import & export
Memories can now be imported and exported by modules. Note that the
multi-memory proposal is not yet standardized, but can be easily
supported with the current architecture.

This features enable running simple Rust programs compiled to Wasm (by
accepting but ignoring tables declaration).

Note that memories are handled the same way functions or globals are,
the code is duplicated in some modules and could benefit from a
refactor.