Add a simple kernel
Add a simple kernel to the repository. The kernel is strongly inspired
from https://os.phil-opp.com, for now it mostly supports interrupts
(timer, page faults...) and VGA.

Note: unfortunately the support for multi-targets in a cargo workspace
is a bit disapointing, I couldn't manage to setup the repo such that a
`cargo test` at the root run all the tests. Instead, each crate must be
built and tested independently.