Add support for allocation
This include creating mappings, frame allocations and a global
allocator.
The global allocator is in fact composed of two allocators:
- A general purpose linked-list allocator (the fallback allocator)
- A more efficient fixed-size allocator (the global allocator)
This enables fast allocation and deallocation with low fragmentation
while keeping the flexibility of the linked-list allocator and limiting
the high allocation & deallocation cost.

The allocators are not completely done yet, in particular the fallback
allocator does not yet merge blocks on deallocation and the fixed-size
allocator allocates new blocks one by one from the fallback allocator
(instead of batching for amortizing the cost & reducing fragmentation).
The overall design should be good enough even for pretty intensive
allocation requirements though.