Prepare compiler & runtime integration
- Migrate the compiler, wasm runtime & collections to no_std.
- Add skeletons of heap & code memory allocators in kernel.

This commit prepares the integration of the compiler & runtime inside
the kernel. This requires migrating all the three libraries (compiler,
wasm runtime & collections) to no_std. There were issues with
cranelift_codegen that is not no_std compatible on the latest published
versions (0.82.X) and master branch.
To port the compiler to no_std, I had to fork cranelift (actually the
whole Wasmtime mono-repo that contains cranelift) and patch it for
no_std support. The fixes mostly consisted in replacing a few `std` by
`core` or `alloc` as appropriate. I might want to upstream the patch at
some point.

On the kernel side, I addes the skeletons for the wasm allocators. Those
requires `ocean-wasm` for the traits to be imported, which thus
exercices the no_std attribute of the library.

All the tests are still passing after the migration to no_std :)