Switch to custom toolchain with std support for bare metal
It appears that Cranelift does in fact _not_ support #[no_std]
environments, for understandable reasons.
This left us with two choices:
- Maintain a fork of Cranelift and potentially all the necessary
  dependencies that do not support #[no_std].
- Support std in the kernel, which requires maintaining a fork of Rust
  itself.

After poking around a bit with the first option, I decided to go for the
second one, for the following reasons:
- There was in fact a lot of crates to fork and patch.
- Cranelift is evolving quickly, meaning that I would need to frequently
  update the forked crates to try out new features.
- There seems to be limited interest in #[no_std] compatibility, see for
  instance the new regaloc2 which just uses std with no opt-out option
  (as of today).
- The rust toolchain is rather stable. New versions comes on a regular
  basis, but compiler versions are usually supported for quite some time
  by most crated.
- The rust tooling is good, it's not that painful to maintain a fork
  (and install a local toolchain with `rustup link`).

This commit migrates the kernel to use std, with a custom toolchain
built from a fork of Rust [1]. For now only the binary is built against
std, but it can successfully be built with a compiler and still passes
all the tests (the binary size is much bigger though, 47Mb! We might be
able to fix that latter through some compiler option magic, like `-Z
build-std`).

[1]: https://github.com/CharlyCst/rust