Renaming kernel to Coral
Prevously the kernel was called Ocean, the explanation behind the name
is that I envisionned a single address-space with isolated memory
regions, like islands in the ocean.
But after thinking about it I don't think it's a good idea to constraint
the design to single address space. Imagine downloading code from the
internet, do we want to run it in the same address space as everything
else? Probably not, software sandboxing can still be subject to bugs.
Moreover, as we control how the code is compilled, we can transparently
change the isolation mechanism! So I defenitely plan to explore that
kinds of things too!
All in all, the name Ocean doesn't make that much sense now, and I like
Coral better :)