Execute first userspace program 🎉
Once the kernel booted correctly, the control is transferred to the
userboot wasm program. For now the program can't do anything meaningful
except returning a value, because the kernel does not expose any
function to the wasm sandbox.

Successfully compiling wasm programs from the kernel required a few
adjustments to Cranelift: the "core" feature was broken, but the "std"
features was not working even with a custom minimal std built with
coral. Instead, a replaced the "core" feature by a "restricted-std"
feature which basically removes the `#![no_std]` but keeps the
`not(feature = "std")` feature gates. With this change, Cranelift can
successfully compile and execute with the Coral std (despite lacking
most advanced std features such as thread local variables).