Add rudimentary support for native function call inside instances
Calling native function from instances is now supported.
A new kind of functions (`Native`) was added, and can be created the
newly introduced `NativeModule`s. In the future NativeModules can be
extended to support other kinds of native resources (e.g. variables or
memories).

For now the support is very limited and unsecure: there is no
typechecker nor mechanism to enforce the expected calling convention for
native functions.