Prototype kernel handles and capabilities
This exposes a new "handle" table to userboot which is initialized with
a single handle refering to a "buffer" object.
For now handles are represented by 64 bits enums, 32 of those are used
for the object index, 8 for the enum tag and the 24 others can be used
for permissions. In case 24 bits of permissions is not enough, we can
add a level of indirection by transforming handles into opaque 32 or 64
bits objects pointing to a vector of actual handles which can be wider
than 64 bits.