Introduce kernel object & rework memory areas
This commit introduces the concept of kernel objects: objects that can
be manipulated from userspace through externref handles.

The first kernel objects are VMAs. Exposing VMAs as kernel object
required a rework of the memory areas and allocator, as the vma must now
be shered with both kernel threads and userspace instances. In particular the
memory area API switched to interior mutability (uses only shared references).
To still enable easy and safe initialization of memory areas, the concept of
"exclusive areas" was added: an area is exclusive after being allocated, until
it is turned into a shared area (by wrapping it into an ARC at the moment). The
API is a bit more verbose, but that's the best I could come up with at the
moment.

Short term future work will be to register instances' memories as kernel
objects, so that they can be manipulated through system calls.