Module instantiation rework
Previously module instantiation were completely handled by the `wasm`
crate, taking an allocator as argument to allocate the necessary memory
areas. This approach did not enable the runtime to take action in
response to heap or table creation for instance, which is unfortunate as
they should be registered as kernel objects by Coral.

This commit introduces a new `Runtime` trait that is implemented by the
wasm embedded (e.g. the Coral kernel) and who is responsible for
performing heap, code and table allocations. This enables Coral to take
actions in response to heap creation, such as registering the heap as a
new kernel object.