Test using a justfile to ease running common commands
Justfiles are just like Makefiles, but without all the build-system and
legacy stuff, they're just aliases for common sequences of commands.
In particular, it is a bit painful to run all the tests as cargo don't
handle the mutiple targets properly (and thus we can't run `cargo test`
at the root). The justfile enables that with a simple `just test`.

See https://github.com/casey/just for details and installation
instructions.