Remove wasm::MemoryAreaAllocator
We no longer need the `wasm::MemoryAreaAllocator` trait as the
`wasm::Runtime` is now responsible for all allocations during instances
lifetime.
This commit removes and simplifies a bunch of traits related to memory
management in the `wasm` crates.

While modifying the kernel VMAs, I also renamed the struct from
`VirtualMemoryArea` to `Vma`. As much as I like explicit name, this one
is pretty long and everyone knows the acronym anyways... All the
documentation should still refers to the full length version so that the
acronym is discoverable to newcomers.