Introduces runtime context for stateful module instantiation
This commits introduces a `Context` type on the `Runtime` trait. A
context is first created during module instantiation, and then passed to
all subsequent methods required for instantiation process (e.g. creating
new heaps).

This enables the runtime to keep track of the currently instantiated
context. Coral uses this feature to store all allocated heaps in the
context and add corresponding handles in the first allocated table. This
allows instances to easily obtain an handle to their own heap.