First "real" syscall: wma_write
This syscall enables an instance to copy bytes from one VMA to another.
In particular it can be used to draw pixels on the screen as the
userboot instance is given a handle to the screen VMA.

Currently the syscall works, but the userboot program seems to be buggy:
it doesn't draw the expected characters nor at the expected position.
Ideally we want to move to Rust-based programs instead of raw
web-assembly anyways, so we will have to re-write it.