Switch to rust-based userboot
This commit switched from the old Wasm hand-written userboot to a new
version written in Rust. This is the first *real* Rust program running
on Coral!

To get there, I had to fix a few bugs:
- Wasm page size is 0x10000, not 0x1000 (64Ki instead of 4Ki).
- The first *externref* table must be filled with the kernel object
  handles.

Note 1: Writing Rust code that manipulate tables is not yet possible, so
this userboot version was obtained by patching a Wasm file produced by
rustc. Once mature enough, the patching toolchain will be included in
this repository.

Note 2: The user code is now responsible for writing to the screen,
therefore the kernel logs should be router elsewhere in a future
version. For now they are just disabled (hence the warnings).