Add userland programs to main repository
This commits port the Rust-based userboot to the main repository. The
program is compiled in two steps: first the rust code is compiled to
wasm32-unknown-unknown, and second the wasm file is patched to use
handles with systemcalls (which can't be manipulated natively from Rust
code).
The patching is done using a new crate, coral-bindgen. A CLI enable
simple patching by providing a configuration file (in the TOML format)
describing which tables to add and how to patch function calls.

To build userland, simply run `just userland`. The newly built files
will be picked up for the next `just run`.