Scheduler refactor
The previous scheduler did not allow for non-blocking task insertion,
which is highly desirable in our kernel (for e.g. scheduling a component
callback after an event dispatch).
The new design is based on Arc to directly enqueue tasks into the
scheduler's main task queue. On wake, the waker simply enqueue the task
itself, which is non-blocking using the crossbeam crate.

With this change, events are dispatched to listeners which are in turn
scheduled to react to the event. This opens the door for more userland
applications, such as keyboard drivers and timers.