Add support for arguments in event dispatcher
Dispatchers can now call instances functions with arguments (at most 5
with the current limitations). This enables porting the keyboard driver
to userland and processing the scancode there.
The side effect of porting the keyboard driver to userland was a
non-negligible incease in userland complexity, I had to increase heap
size and optimization level so that the kernel can compile userland in a
reasonable amount of time without running into an OOM error.