Syscall 2.0
Major rewrite of syscall API & ABI.
This rewrite was driven by the need to return more than one value for
some syscalls, which can't be handled by the standard SystemV calling
convention used within Coral (SystemV can return struct, but the ABI is
complicated...). Coral implements the same ABI than Wasmtime, that is if
there is multiple returns the caller pre-allocate space for the return
values beyond the first and pass a pointer to that space as argument.
A huge part of the work is to abstract this ABI, this is done through
a macro and a set of traits representing the different WebAssembly types
and their corresponding ABI. For now the macro is not really ergonomic,
as the user has to repeat all the function's types, in the future this
can be fixed using a procedural macro instead, but this is good enough
for now.

There are still some bug though, as demonstrated by the new failing
test! I'll work on a fix next.